Moved folder context out of assets
