+++ src/main.jsx
@@ -1,7 +1,7 @@
 import React from 'react'
 import ReactDOM from 'react-dom/client'
 import { BrowserRouter as Router } from 'react-router-dom';
-import AuthContextProvider from './assets/context/AuthContext';
+import AuthContextProvider from '../src/context/AuthContext';
 import './index.css';
 import App from './App';
